wxpy3.0 compat
Update for wxPython 3.0
wx.InitAllImageHandlers() issues a deprecation warning with wxPython 3.0
and is a no-op with wxPython 2.8.
Deprecated constants like wxSAVE are gone from the C++ API, so wx.SAVE, etc
are also likely to go from wxPython soon.
The alias "wx.Color" is gone in wxPython 3.0 - only "wx.Colour" is now
supported, so update docstring to use the correct class name.
Author: Olly Betts <olly@survex.com>
Last-Update: 2014-08-21